backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# File Processing
pdfplumber==0.10.3
pdf2image==1.16.3
python-docx==0.8.11
openpyxl==3.11.0
pillow==10.1.0
pytesseract==0.3.10

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
aiofiles==23.2.1
uuid6==1.0.3

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
